+++ src/main/java/com/megazone/goop/racingcar/car/CarName.java
@@ -1,4 +1,4 @@
-package com.megazone.goop.racingcar.logic;
+package com.megazone.goop.racingcar.car;
 
 import java.text.MessageFormat;
