--- src/main/java/com/megazone/goop/racingcar/logic/RoundRecord.java
@@ -1,6 +1,0 @@
-package com.megazone.goop.racingcar.logic;
-
-import java.util.List;
-
-public record RoundRecord(List<CarDistance> carDistances) {
-}
