+++ src/main/java/com/megazone/goop/racingcar/race/RacePlan.java
@@ -1,6 +1,8 @@
-package com.megazone.goop.racingcar.logic;
+package com.megazone.goop.racingcar.race;
 
 import java.util.List;
+
+import com.megazone.goop.racingcar.car.CarName;
 
 public record RacePlan(
     List<CarName> CarNames,
