+++ src/main/java/com/megazone/goop/racingcar/race/RoundRecord.java
@@ -1,0 +1,8 @@
+package com.megazone.goop.racingcar.race;
+
+import java.util.List;
+
+import com.megazone.goop.racingcar.car.CarDistance;
+
+public record RoundRecord(List<CarDistance> carDistances) {
+}
