+++ src/main/java/com/megazone/goop/racingcar/ui/InputView.java
@@ -2,8 +2,8 @@ package com.megazone.goop.racingcar.ui;
 
 import java.util.*;
 
-import com.megazone.goop.racingcar.logic.CarName;
-import com.megazone.goop.racingcar.logic.RacePlan;
+import com.megazone.goop.racingcar.car.CarName;
+import com.megazone.goop.racingcar.race.RacePlan;
 
 public class InputView {
