+++ src/main/java/com/megazone/goop/racingcar/util/RandomGenerator.java
@@ -1,4 +1,4 @@
-package com.megazone.goop.racingcar.logic;
+package com.megazone.goop.racingcar.util;
 
 import java.util.Random;
