+++ src/test/java/com/megazone/goop/racingcar/util/StringSumCalculatorTest.java
@@ -1,9 +1,11 @@
-package com.megazone.goop.racingcar;
+package com.megazone.goop.racingcar.util;
 
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+
+import com.megazone.goop.racingcar.util.StringSumCalculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
